optimize solution
When we have multiple, (4, 27) for example, we can directly go to
(4, 3).
So the test case was failing because of that, hard to know because
there was no hint regarding the fact that it was because of slow
computation and not because of correctness of the algo :/

Anyway as there was no more corner cases, with the 2 added tests for
the "0" cases, it was the only possible reason for a failure.