first version of the algorithm for 4.1
We are doing a DFS starting with all exits, we are looking for path,
and everytime we are exiting a node we are keeping track of the
remaining flow that could have been taken from this node, like that
if we are visiting again this node, we will try to get the remaining
flow if some is available.
The recursive method is returning the flow that we took.