response for challenge 4.2
First we will look for a negative cycle using Bellman Ford
algorithm. A negative cycle would mean we can get as much time as
we need to do whatever path.

Then we will do a brute force algorithm to find all paths. We are
cutting the paths as soon as they are taking too much time, or if we
visit a same vertex with the same bunnies, but with a worst time.

In order to optimize a bit the memory and the replication of data,
we use bits in an int to store a solution, as there can be only 5 bunnies, so a solution
fit in 5 bits.