initial algo for challenge 5
The algo is working for second case and third case, but it is
failing for the first case.
It will be hard to debug it :/

It is also failing 3 of four hidden cases, so it seems that there is
a glitch in the algo :(

Also when computing the possible combinations, there is clearly a
vertical symmetry (if we remove the constraint from left, but do we
need that before merging with previous row?). So we could probably
leverage that, by analyzing only 8 pieces instead of 16.
And then we could also improve a bit the algo with some memoization,
we do have up to 7 rows, but up to 5 columns, so there is some
chances to have repeating columns.